requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
flask_cors
MySQL.connector
mongoengine
neo4j
redis
PyMySQL
bcrypt
PyJWT
pymongo
